added bad request validation

+++ app/Http/Controllers/Api/Setting/SettingController.php
@@ -19,6 +19,9 @@ class SettingController extends AppBaseController
 
     public function updateBasedOnActivity($activity,UpdateSettingRequest $request){
         try {
+            if (!in_array($activity, ['image', 'account', 'password','privacy','notification'])) {
+                return $this->sendError(__('responses.handler_bad_request'), 400);
+            }
             switch ($activity){
                 case 'image':
                     $removeProfile = $this->settingRepository->removeProfileImage();
